+++ components/navigation/TabNavigation.jsx
@@ -1,0 +1,12 @@
+import { View, Text } from 'react-native';
+import React from 'react';
+
+const TabNavigation = () => {
+  return (
+    <View>
+      <Text>TabNavigation</Text>
+    </View>
+  );
+};
+
+export default TabNavigation;
