+++ screens/index.js
@@ -2,10 +2,11 @@ import LoginScreen from "./AuthScreen/LoginScreen";
 import HomeScreen from "./HomeScreen";
 import MapScreen from "./MapScreen";
 import SignUpScreen from "./AuthScreen/SignUpScreen";
-
+import WelcomeScreen from './WelcomeScreen';
 export {
     HomeScreen,
     MapScreen,
     LoginScreen,
     SignUpScreen,
+    WelcomeScreen,
 }
